requirements.txt
google-search-results==2.4.2
pandas==2.2.3
streamlit==1.35.0
googlesearch-python
matplotlib
python-dotenv